refactor: codestyle
