+++ app/service_10k/app/entities.py
@@ -1,5 +1,6 @@
 from pydantic import BaseModel
 
+
 class PredictResponse(BaseModel):
     negative_probability: float
     positive_probability: float
