+++ app/service_10k/app/service.py
@@ -1,32 +1,44 @@
-# import app.service_10k.app.entities as entities
-import entities
-# import app.service_10k.app.config as config
-import config
+import logging
 import torch
 import joblib
 from transformers import DistilBertTokenizer, DistilBertModel
-import logging
+
+from app.service_10k.app import entities
+from app.service_10k.app import config
 
 logging.basicConfig(level=config.LOGGING_LEVEL)
 
 
 class ModelContext:
     def __init__(self):
+        """
+        Initialization of context
+        """
         self.model_lr = None
         self.tokenizer = None
         self.model_bert = None
         self.device = torch.device(config.DEVICE)
 
     def load_models(self):
+        """
+        Loading model context
+        """
         self.model_lr = joblib.load(config.PATH_LR_MODEL)
         self.tokenizer = DistilBertTokenizer.from_pretrained('distilbert-base-uncased')
         self.model_bert = DistilBertModel.from_pretrained('distilbert-base-uncased')
 
+
 model_context = ModelContext()
 
+
 def predict_text(report: str) -> entities.PredictResponse:
+    """
+    This method involves tokenization and vectorization of text, as well as forecasting
+    """
     logging.info('Токенизация начата')
-    tokens = model_context.tokenizer(report[42080:47080], return_tensors='pt', padding=True, truncation=True).to(model_context.device)
+    tokens = model_context.tokenizer(report[42080:47080],
+                                     return_tensors='pt',
+                                     padding=True, truncation=True).to(model_context.device)
     logging.info('Токенизация завершена')
     logging.info('Старт расчета эмбеддингов')
     with torch.no_grad():
